Move room service tests to separate folder
Wasn't going to separate the tests at first but I'm noticing how there won't be any order once I add tests for the other entity services.
