+++ ShapeDungeon.Tests/ServiceTests/Rooms/GetRoomServiceTests.cs
@@ -7,7 +7,7 @@ using ShapeDungeon.Repos;
 using ShapeDungeon.Services.Rooms;
 using System.Threading.Tasks;
 
-namespace ShapeDungeon.Tests.ServiceTests
+namespace ShapeDungeon.Tests.ServiceTests.Rooms
 {
     internal class GetRoomServiceTests
     {
